+++ api/app/config.py
@@ -1,3 +1,5 @@
+from functools import lru_cache
+
 from pydantic import BaseSettings, Field
 
 
@@ -8,3 +10,9 @@ class Settings(BaseSettings):
     class Config:
         env_file = ".env"
         env_file_encoding = "utf-8"
+
+
+@lru_cache
+def get_settings():
+    """dependency for returning settings"""
+    return Settings()
